renaming visualization manager
