+++ src/neuroconnect/visualization_manager.py
